refactor: extract trade footer
